+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/NPC/EffectSpawnRandomL4D2Companion.lua
@@ -13,7 +13,9 @@ local function MakeTint(r, g, b)
     }
 end
 
+---@type table<integer, string>
 local weapons = { "Base.Axe", "Base.BaseballBat", "Base.BaseballBat_Nails", "Base.Plank_Nails" }
+local special_weapon = "Base.Shotgun"
 
 ---@param zombie IsoZombie
 ---@param effectName string
@@ -161,10 +163,18 @@ function EffectSpawnRandomL4D2Companion:OnStart()
     npc:initializeHuman()
     npc.npcGroup = ChaosNPCGroupID.COMPANIONS
 
-    local weaponIndex = ChaosUtils.RandArrayIndex(weapons)
-    local weapon = weapons[weaponIndex]
+    local weapon = ""
 
-    if weapon then
+
+    if ChaosUtils.RandFloat(0.0, 100.0) < 25.0 then
+        weapon = special_weapon
+    else
+        local weaponIndex = ChaosUtils.RandArrayIndex(weapons)
+        weapon = weapons[weaponIndex]
+    end
+
+
+    if weapon and weapon ~= "" then
         npc:SetWeapon(weapon)
     end