+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/NPC/EffectTameCaveman.lua
@@ -1,0 +1,60 @@
+---@class EffectTameCaveman : ChaosEffectBase
+EffectTameCaveman = ChaosEffectBase:derive("EffectTameCaveman", "tame_caveman")
+
+function EffectTameCaveman:OnStart()
+    ChaosEffectBase:OnStart()
+    print("[EffectTameCaveman] OnStart" .. tostring(self.effectId))
+    local player = getPlayer()
+    if not player then return end
+
+    local playerSquare = player:getSquare()
+    if not playerSquare then return end
+
+    local randomSquare = ChaosPlayer.GetRandomSquareAroundPlayer(player, nil, 2, 4, 50, true, true, false)
+    if not randomSquare then return end
+
+    local newZombies = ChaosZombie.SpawnZombieAt(
+        randomSquare:getX(),
+        randomSquare:getY(),
+        randomSquare:getZ(),
+        1,
+        "Naked",
+        0
+    )
+
+    local zombie = newZombies and newZombies:getFirst() or nil
+    if not zombie then return end
+
+    local npc = ChaosNPC:new(zombie, self.effectNickname)
+    npc:SetHealthGroup(CHAOS_NPC_HEALTH_GROUP.STRONG)
+    npc:initializeHuman()
+    npc.npcGroup = ChaosNPCGroupID.COMPANIONS
+
+    --- Hair / beard
+    local humanVisual = zombie:getHumanVisual()
+    humanVisual:setHairModel("LongBraids02")
+    humanVisual:setBeardModel("Long")
+    local hairColor = ImmutableColor.new(89 / 255, 56 / 255, 30 / 255)
+    humanVisual:setHairColor(hairColor)
+    humanVisual:setNaturalHairColor(hairColor)
+    humanVisual:setBeardColor(hairColor)
+    humanVisual:setNaturalBeardColor(hairColor)
+
+    --- Remove default clothes
+    zombie:getItemVisuals():clear()
+    zombie:clearWornItems()
+
+    --- Add new clothes
+    ChaosZombie.AddZombieClothesBatch(zombie, {
+        { type = "Base.Underpants_Hide" },
+        { type = "Base.Vest_Hide" },
+    })
+
+    zombie:resetModelNextFrame()
+
+    npc:SetWeapon("Base.LargeBoneClub")
+end
+
+function EffectTameCaveman:OnEnd()
+    ChaosEffectBase:OnEnd()
+end